Update Integrations Page to Link to Python and NodeJS Libraries
This is a freshening of the API Integrations documentation page to link to the official Python and NodeJS client libraries maintained by BTCPay.

I recently merged a few changes into the Python library and uploaded the official BTCPay python client to PyPI so that it can be installed by Python developers using the pip package manager. It appears that the NodeJS guys have done the same thing with npm for their library.

I did reference the PHP and Ruby libraries in the doc as well, however when I looked through those two, it appears that BTCPay's forks to the Bitpay versions are not uploaded to the PHP and Ruby repositories. That would mean that a user following the install instructions in their READMEs using Ruby Gems or PHP Composer would be installing the unmodified Bitpay version rather than the modified BTCPay version. If anyone out there knows Ruby or PHP, getting the BTCPay forks into the repos and updating the READMEs might be an easy yet valuable project. Unfortunately I don't have any familiarity with Ruby or PHP, so my interpretation of the above could be wrong!