Update Troubleshooting.md
I propose to change the suggestion to get a log output, from:
`docker logs --tail 40`
to
`docker logs --tail 100` 
The reason is simple. As a newbie (as i am) you probably just copy and paste, and with `--tail 40` there is a non negligible chance that you miss the key information. This happened to me.