Adds info on preinstalled hardware
Adds info on preinstalled harware boxes.
Relates to #515

I name the document and the title "Hardware as A Service" because I like the sound of it.
Name suggestions appreciated.
DIYNodes for example.

Feel free to comment on wording, missing info etc ...

I will commit the summary change in config.js once this PR is mergeable.