Update linkcheck and GH Actions pipeline
The linkcheck is now done locally against the current build. The pipeline also runs for pull request, so that we can ensure we do not break anything with those.
Deployment still happens only for `master` and if linkcheck succeeds. Though I also added an exclude list for external URLs that do not respond properly, this might be too restrictive. We shall see.