Add Light POS section under Point of Sale in apps.md
Adds Light POS section under Point of Sale in apps.md

Closes #735

Specifically
* Adds a paragraphe mentionning the three possible views for a POS app
* Adds some `code` formatting for easier reading
* Adds a link to the blog post relating to the new Light POS feature.
Be advised, I updated said blog post to include a custom anchor to link directly to the relevant paragraphe `Simple Point Of Sale`.
Said link can be deleted in the future when we feel that most users will have updated to that version.