handling for umount error
umount /dev/sda1

If the drive is in use, there will be an unhandled error, the drive won't unmount and can't be prepared as intended. The script will take a long time and the user will find out about the failure at the end.

To handle this, a simple OR operator with an error block can exit the script. Might be worth adding this drive preparation before updates and docker installation to get any such error out of the way early.

The proposed adjustment is:

umount /dev/sda1 || { echo -e "\nUnable to unmount /dev/sda1 - Please make sure the device is not busy and try again" ; exit ; }

and again for the nvme drive.