error handline of failed unmount
For sda1 and nvme0n1p1

If the device fails to unmount, say due to a busy device, the code will get to the end but will fail. The user has to wait a long time to get the failed feedback. Better to exit immediately. 

SUGGESTION...
It might be worth doing the error check earlier by moving the unmounting procedure before apt update/upgrade and before docker install.