Add Subscriptions.md
This PR adds Subscription.md in preparation for the upcoming 2.3.0 release. It uses PR description that @NicolasDorier provided and all the videos he made as a starting point. I leveraged AI for this one, but it was done in a responsible way, meaning data set was strictly controlled and I've done at least 2 rounds of reviews and deployment locally before shipping it. I've tested it locally to ensure everything works, but would appreciate reviews here for sanity check.

fix img path capitalizaiton causing CI to fail